requirements.txt
accelerate==1.10.0
diffusers==0.35.1
ftfy==6.3.1
huggingface-hub==0.34.4
numpy==2.1.2
packaging==25.0
Pillow==11.0.0
psutil==7.0.0
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scipy==1.16.1
tokenizers==0.21.4
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.55.4
typing-extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.13
zipp==3.23.0
